requirements.txt
PyPDF2
dialogflow 
# Add Flask if you're building a web frontend
